+++ sdks/typescript/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to Hatchet's TypeScript SDK will be documented in this chang
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.28.2] - 2026-08-05
+
+- Improve gRPC error logging when hitting max message size limits on `putStream` calls.
+
 ## [1.28.1] - 2026-07-30
 
 ### Fixed